LayoutTests/storage/domstorage/localstorage/storagetracker/storage-tracker-7-usage-expected.txt

Test API that reports local storage disk usage by origin. NOTE: this test DEPENDS on storage-tracker-6-create.html to create local storage for the 'file://' origin. If this test runs in a separate DRT instance from storage-tracker-6-create.html, it may FAIL if there is no local storage!

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testRunner.originsWithLocalStorage().toString() is "file__0"
PASS testRunner.localStorageDiskUsageForOrigin('file://') > 6100 is true
PASS successfullyParsed is true

TEST COMPLETE
